refactor(zk-explorer): Move dashboard tabs into its own component
